+++ src-core/library/index.ts
@@ -5,7 +5,6 @@
  */
 
 export { Library } from './library';
-export type { ImportJob, ImportJobEvents, ImportProgress } from './library';
+export type { ImportJob, ImportJobEvents, ImportProgress } from './import-job';
 export type { LibraryOptions } from './library';
-export { TrackImporter } from './track-importer';
 export type { Track } from './track';
